Move splitCommitTransactionRequest and shouldSplitCommitTransactionRequest to NativeAPI.actor.cpp
