+++ fdbclient/CMakeLists.txt
@@ -33,7 +33,6 @@ set(FDBCLIENT_SRCS
   Knobs.h
   ManagementAPI.actor.cpp
   ManagementAPI.actor.h
-  MasterProxyInterface.cpp
   MasterProxyInterface.h
   MetricLogger.actor.cpp
   MetricLogger.h
